Add more detail to examples in Readme.md
When following the examples, I didn't realise that the Authorize Controller example was different to the first Client Credentials example. I also made the mistake of adding the additional grant type at the end of token.php, rather than in the same place as the previous one. This patch clarifies the readme file.