Add example for testing with Google OAuth2 Playground.
Having followed through the examples in the Readme for creating a local
controller, I wanted to test it out for real. I eventually stumbled on the
Google Playground, but it still took me a while to work out how to get it
working for real. I think this example would be useful to include for
others.