IP verification
@see https://tools.ietf.org/html/rfc7662#section-2.1

Check requester's requester IP if required to match.

If the requester passed the IP of its own requester, we must check that this IP is in the subnet of the client application identified by 'aud'.

We parse the client redirect URI to get the host and we obtain client IPs from DNS.
Then we compare the requester IP to the client IPs using the mask defined in server configuration.